internal/services/datalake/sdk/datalakeanalytics/2016-11-01/datalakestoreaccounts/id_datalakestoreaccount.go
package datalakestoreaccounts

import (
	"fmt"
	"strings"

	"github.com/hashicorp/go-azure-helpers/resourcemanager/resourceids"
)

type DataLakeStoreAccountId struct {
	SubscriptionId string
	ResourceGroup  string
	AccountName    string
	Name           string
}

func NewDataLakeStoreAccountID(subscriptionId, resourceGroup, accountName, name string) DataLakeStoreAccountId {
	return DataLakeStoreAccountId{
		SubscriptionId: subscriptionId,
		ResourceGroup:  resourceGroup,
		AccountName:    accountName,
		Name:           name,
	}
}

func (id DataLakeStoreAccountId) String() string {
	segments := []string{
		fmt.Sprintf("Name %q", id.Name),
		fmt.Sprintf("Account Name %q", id.AccountName),
		fmt.Sprintf("Resource Group %q", id.ResourceGroup),
	}
	segmentsStr := strings.Join(segments, " / ")
	return fmt.Sprintf("%s: (%s)", "Data Lake Store Account", segmentsStr)
}

func (id DataLakeStoreAccountId) ID() string {
	fmtString := "/subscriptions/%s/resourceGroups/%s/providers/Microsoft.DataLakeAnalytics/accounts/%s/dataLakeStoreAccounts/%s"
	return fmt.Sprintf(fmtString, id.SubscriptionId, id.ResourceGroup, id.AccountName, id.Name)
}

// ParseDataLakeStoreAccountID parses a DataLakeStoreAccount ID into an DataLakeStoreAccountId struct
func ParseDataLakeStoreAccountID(input string) (*DataLakeStoreAccountId, error) {
	id, err := resourceids.ParseAzureResourceID(input)
	if err != nil {
		return nil, err
	}

	resourceId := DataLakeStoreAccountId{
		SubscriptionId: id.SubscriptionID,
		ResourceGroup:  id.ResourceGroup,
	}

	if resourceId.SubscriptionId == "" {
		return nil, fmt.Errorf("ID was missing the 'subscriptions' element")
	}

	if resourceId.ResourceGroup == "" {
		return nil, fmt.Errorf("ID was missing the 'resourceGroups' element")
	}

	if resourceId.AccountName, err = id.PopSegment("accounts"); err != nil {
		return nil, err
	}
	if resourceId.Name, err = id.PopSegment("dataLakeStoreAccounts"); err != nil {
		return nil, err
	}

	if err := id.ValidateNoEmptySegments(input); err != nil {
		return nil, err
	}

	return &resourceId, nil
}

// ParseDataLakeStoreAccountIDInsensitively parses an DataLakeStoreAccount ID into an DataLakeStoreAccountId struct, insensitively
// This should only be used to parse an ID for rewriting to a consistent casing,
// the ParseDataLakeStoreAccountID method should be used instead for validation etc.
func ParseDataLakeStoreAccountIDInsensitively(input string) (*DataLakeStoreAccountId, error) {
	id, err := resourceids.ParseAzureResourceID(input)
	if err != nil {
		return nil, err
	}

	resourceId := DataLakeStoreAccountId{
		SubscriptionId: id.SubscriptionID,
		ResourceGroup:  id.ResourceGroup,
	}

	if resourceId.SubscriptionId == "" {
		return nil, fmt.Errorf("ID was missing the 'subscriptions' element")
	}

	if resourceId.ResourceGroup == "" {
		return nil, fmt.Errorf("ID was missing the 'resourceGroups' element")
	}

	// find the correct casing for the 'accounts' segment
	accountsKey := "accounts"
	for key := range id.Path {
		if strings.EqualFold(key, accountsKey) {
			accountsKey = key
			break
		}
	}
	if resourceId.AccountName, err = id.PopSegment(accountsKey); err != nil {
		return nil, err
	}

	// find the correct casing for the 'dataLakeStoreAccounts' segment
	dataLakeStoreAccountsKey := "dataLakeStoreAccounts"
	for key := range id.Path {
		if strings.EqualFold(key, dataLakeStoreAccountsKey) {
			dataLakeStoreAccountsKey = key
			break
		}
	}
	if resourceId.Name, err = id.PopSegment(dataLakeStoreAccountsKey); err != nil {
		return nil, err
	}

	if err := id.ValidateNoEmptySegments(input); err != nil {
		return nil, err
	}

	return &resourceId, nil
}
